+++ simapp/sim_test.go
@@ -21,6 +21,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	vesting "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
 	authzkeeper "github.com/cosmos/cosmos-sdk/x/authz/keeper"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
@@ -36,6 +37,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/NibiruChain/nibiru/v2/app"
+	"github.com/NibiruChain/nibiru/v2/app/codec"
 	devgastypes "github.com/NibiruChain/nibiru/v2/x/devgas/v1/types"
 	epochstypes "github.com/NibiruChain/nibiru/v2/x/epochs/types"
 	inflationtypes "github.com/NibiruChain/nibiru/v2/x/inflation/types"
@@ -59,6 +61,17 @@ type StoreKeysPrefixes struct {
 	Prefixes [][]byte
 }
 
+// makeEncodingConfig, similar to [app.MakeEncodingConfig], creates an
+// EncodingConfig for an amino based test configuration. However, this function
+// registers interfaces and types that are expected by default in the Cosmos-SDK
+// even if they are disabled on Nibiru. This is the case for x/vesting Cosmos-SDK
+// module.
+func makeEncodingConfig() codec.EncodingConfig {
+	encCfg := app.MakeEncodingConfig()
+	vesting.RegisterInterfaces(encCfg.InterfaceRegistry)
+	return encCfg
+}
+
 func TestFullAppSimulation(t *testing.T) {
 	config := simcli.NewConfigFromFlags()
 	config.ChainID = SimAppChainID
@@ -78,7 +91,7 @@ func TestFullAppSimulation(t *testing.T) {
 	appOptions[flags.FlagHome] = app.DefaultNodeHome
 	appOptions[server.FlagInvCheckPeriod] = simcli.FlagPeriodValue
 
-	encoding := app.MakeEncodingConfig()
+	encoding := makeEncodingConfig()
 	app := app.NewNibiruApp(logger, db, nil, true, encoding, appOptions, baseapp.SetChainID(SimAppChainID))
 	require.Equal(t, "Nibiru", app.Name())
 	appCodec := app.AppCodec()
@@ -132,7 +145,7 @@ func TestAppStateDeterminism(t *testing.T) {
 		for j := 0; j < numTimesToRunPerSeed; j++ {
 			db := dbm.NewMemDB()
 			logger := log.NewNopLogger()
-			encoding := app.MakeEncodingConfig()
+			encoding := makeEncodingConfig()
 
 			app := app.NewNibiruApp(logger, db, nil, true, encoding, appOptions, baseapp.SetChainID(SimAppChainID))
 			appCodec := app.AppCodec()
@@ -191,7 +204,7 @@ func TestAppImportExport(t *testing.T) {
 	appOptions[flags.FlagHome] = app.DefaultNodeHome
 	appOptions[server.FlagInvCheckPeriod] = simcli.FlagPeriodValue
 
-	encoding := app.MakeEncodingConfig()
+	encoding := makeEncodingConfig()
 	oldApp := app.NewNibiruApp(logger, db, nil, true, encoding, appOptions, baseapp.SetChainID(SimAppChainID))
 	require.Equal(t, "Nibiru", oldApp.Name())
 	appCodec := oldApp.AppCodec()
@@ -315,7 +328,7 @@ func TestAppSimulationAfterImport(t *testing.T) {
 	appOptions[flags.FlagHome] = app.DefaultNodeHome
 	appOptions[server.FlagInvCheckPeriod] = simcli.FlagPeriodValue
 
-	encoding := app.MakeEncodingConfig()
+	encoding := makeEncodingConfig()
 	oldApp := app.NewNibiruApp(logger, db, nil, true, encoding, appOptions, baseapp.SetChainID(SimAppChainID))
 	require.Equal(t, "Nibiru", oldApp.Name())
 	appCodec := oldApp.AppCodec()